Changes to DUAL channel plugins + rename
