+++ Makefile
@@ -5,28 +5,28 @@ DESTDIR =
 all: plugins
 
 plugins:
-	$(MAKE) -C midi-switchbox-1inx2out.lv2
-	$(MAKE) -C midi-switchbox-1inx3out.lv2
-	$(MAKE) -C midi-switchbox-2inx1out.lv2
-	$(MAKE) -C midi-switchbox-3inx1out.lv2
-	$(MAKE) -C midi-switchbox-2inx4out.lv2
-	$(MAKE) -C midi-switchbox-4inx2out.lv2
+	$(MAKE) -C midi-switchbox_1-2.lv2
+	$(MAKE) -C midi-switchbox_1-3.lv2
+	$(MAKE) -C midi-switchbox_2-1.lv2
+	$(MAKE) -C midi-switchbox_3-1.lv2
+	$(MAKE) -C midi-switchbox_1-2_2C.lv2
+	$(MAKE) -C midi-switchbox_2-1_2C.lv2
 	$(MAKE) -C peak-to-cc.lv2
 
 install:
-	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox-1inx2out.lv2
-	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox-1inx3out.lv2
-	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox-2inx1out.lv2
-	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox-3inx1out.lv2
-	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox-2inx4out.lv2
-	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox-4inx2out.lv2
+	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox_1-2.lv2
+	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox_1-3.lv2
+	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox_2-1.lv2
+	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox_3-1.lv2
+	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox_1-2_2C.lv2
+	$(MAKE) install PREFIX=$(PREFIX) -C midi-switchbox_2-1_2C.lv2
 	$(MAKE) install PREFIX=$(PREFIX) -C peak-to-cc.lv2
 
 clean:
-	$(MAKE) clean -C midi-switchbox-1inx2out.lv2
-	$(MAKE) clean -C midi-switchbox-1inx3out.lv2
-	$(MAKE) clean -C midi-switchbox-2inx1out.lv2
-	$(MAKE) clean -C midi-switchbox-3inx1out.lv2
-	$(MAKE) clean -C midi-switchbox-2inx4out.lv2
-	$(MAKE) clean -C midi-switchbox-4inx2out.lv2
+	$(MAKE) clean -C midi-switchbox_1-2.lv2
+	$(MAKE) clean -C midi-switchbox_1-3.lv2
+	$(MAKE) clean -C midi-switchbox_2-1.lv2
+	$(MAKE) clean -C midi-switchbox_3-1.lv2
+	$(MAKE) clean -C midi-switchbox_1-2_2C.lv2
+	$(MAKE) clean -C midi-switchbox_2-1_2C.lv2
 	$(MAKE) clean -C peak-to-cc.lv2
