--- midi-switchbox-1inx3out.lv2/manifest.ttl
@@ -1,7 +1,0 @@
-@prefix lv2:  <http://lv2plug.in/ns/lv2core#> .
-@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
-
-<http://moddevices.com/plugins/mod-devel/midi-switchbox-1inx3out>
-    a lv2:Plugin ;
-    lv2:binary <midi-switchbox-1inx3out.so>  ;
-    rdfs:seeAlso <midi-switchbox-1inx3out.ttl> , <modgui.ttl> .
