+++ midi-switchbox_1-2.lv2/manifest.ttl
@@ -3,6 +3,6 @@
 
 <http://moddevices.com/plugins/mod-devel/MIDI-Switchbox-1inx2out>
     a lv2:Plugin ;
-    lv2:binary <midi-switchbox-1inx2out.so>  ;
-    rdfs:seeAlso <midi-switchbox-1inx2out.ttl> ,
+    lv2:binary <midi-switchbox_1-2.so>  ;
+    rdfs:seeAlso <midi-switchbox_1-2.ttl> ,
                  <modgui.ttl> .
