+++ midi-switchbox_1-2_2C.lv2/midi-switchbox_1-2_2C.c
@@ -7,20 +7,23 @@
 
 #include <stdbool.h>
 #include <stdlib.h>
+#include <stdio.h>
 
 typedef enum {
     PORT_CONTROL_TARGET = 0,
-    PORT_ATOM_IN1,
-    PORT_ATOM_IN2,
-    PORT_ATOM_OUT
+    PORT_MIDI_IN1,
+    PORT_MIDI_IN2,
+    PORT_MIDI_OUT1,
+    PORT_MIDI_OUT2,
+    PORT_MIDI_OUT3,
+    PORT_MIDI_OUT4
 } PortEnum;
 
 typedef struct {
     // keep track of active notes
     bool active_notes[16*127];
 
-    // was last run on the 2nd output?
-    bool was_second;
+	int prev_target;
 
     // URIDs
     LV2_URID urid_midiEvent;
@@ -31,7 +34,10 @@ typedef struct {
     // atom ports
     const LV2_Atom_Sequence* port_events_in1;
     const LV2_Atom_Sequence* port_events_in2;
-    LV2_Atom_Sequence* port_events_out;
+    LV2_Atom_Sequence* port_events_out1;
+    LV2_Atom_Sequence* port_events_out2;
+    LV2_Atom_Sequence* port_events_out3;
+    LV2_Atom_Sequence* port_events_out4;
 } Data;
 
 // Struct for a 3 byte MIDI event
@@ -64,6 +70,8 @@ static LV2_Handle instantiate(const LV2_Descriptor*     descriptor,
     // Map URIs
     self->urid_midiEvent = map->map(map->handle, LV2_MIDI__MidiEvent);
 
+	self->prev_target = 0;
+
     return self;
 }
 
@@ -76,14 +84,23 @@ static void connect_port(LV2_Handle instance, uint32_t port, void* data)
     case PORT_CONTROL_TARGET:
             self->port_target = (const float*)data;
             break;
-    case PORT_ATOM_IN1:
+    case PORT_MIDI_IN1:
             self->port_events_in1 = (const LV2_Atom_Sequence*)data;
             break;
-    case PORT_ATOM_IN2:
-            self->port_events_in2 = (LV2_Atom_Sequence*)data;
+    case PORT_MIDI_IN2:
+            self->port_events_in2 = (const LV2_Atom_Sequence*)data;
             break;
-    case PORT_ATOM_OUT:
-            self->port_events_out = (LV2_Atom_Sequence*)data;
+    case PORT_MIDI_OUT1:
+            self->port_events_out1 = (LV2_Atom_Sequence*)data;
+            break;
+    case PORT_MIDI_OUT2:
+            self->port_events_out2 = (LV2_Atom_Sequence*)data;
+            break;
+    case PORT_MIDI_OUT3:
+            self->port_events_out3 = (LV2_Atom_Sequence*)data;
+            break;
+    case PORT_MIDI_OUT4:
+            self->port_events_out4 = (LV2_Atom_Sequence*)data;
             break;
     }
 }
@@ -99,20 +116,27 @@ static void run(LV2_Handle instance, uint32_t sample_count)
 {
     Data* self = (Data*)instance;
 
-    const bool target_second = (*self->port_target) > 0.5f;
+    const int target = (int)(*self->port_target);
 
     // Get the capacity
-    const uint32_t out_capacity = self->port_events_out->atom.size;
-
+    const uint32_t out_capacity1 = self->port_events_out1->atom.size;
+    const uint32_t out_capacity2 = self->port_events_out2->atom.size;
+    const uint32_t out_capacity3 = self->port_events_out1->atom.size;
+    const uint32_t out_capacity4 = self->port_events_out2->atom.size;
     // Write an empty Sequence header to the output
-    lv2_atom_sequence_clear(self->port_events_out);
+    lv2_atom_sequence_clear(self->port_events_out1);
+    lv2_atom_sequence_clear(self->port_events_out2);
+    lv2_atom_sequence_clear(self->port_events_out3);
+    lv2_atom_sequence_clear(self->port_events_out4);
 
     // LV2 is so nice...
-    self->port_events_out->atom.type = self->port_events_in1->atom.type;
-    self->port_events_out->atom.type = self->port_events_in2->atom.type;
+    self->port_events_out1->atom.type = self->port_events_in1->atom.type;
+    self->port_events_out2->atom.type = self->port_events_in2->atom.type;
+    self->port_events_out3->atom.type = self->port_events_in1->atom.type;
+    self->port_events_out4->atom.type = self->port_events_in2->atom.type;
 
     // Send note-offs if target port changed
-    if (self->was_second != target_second)
+    if (self->prev_target != target)
     {
         LV2_Atom_MIDI msg;
         memset(&msg, 0, sizeof(LV2_Atom_MIDI));
@@ -130,14 +154,26 @@ static void run(LV2_Handle instance, uint32_t sample_count)
 
                     msg.msg[0] = LV2_MIDI_MSG_NOTE_OFF + i;
                     msg.msg[1] = j;
-                    lv2_atom_sequence_append_event(self->port_events_out,
-                                                            out_capacity,
+
+					printf("send note offs!\n");
+
+                    lv2_atom_sequence_append_event(self->port_events_out1,
+                                                            out_capacity1,
+                                                  (LV2_Atom_Event*)&msg);
+                    lv2_atom_sequence_append_event(self->port_events_out2,
+                                                            out_capacity2,
+                                                  (LV2_Atom_Event*)&msg);
+                    lv2_atom_sequence_append_event(self->port_events_out3,
+                                                            out_capacity3,
+                                                  (LV2_Atom_Event*)&msg);
+                    lv2_atom_sequence_append_event(self->port_events_out4,
+                                                            out_capacity4,
                                                   (LV2_Atom_Event*)&msg);
                 }
             }
         }
 
-        self->was_second = target_second;
+        self->prev_target = target;
     }
 
     // Read incoming events
@@ -163,13 +199,21 @@ static void run(LV2_Handle instance, uint32_t sample_count)
             }
         }
 
-        if (!target_second)
-        {
-            lv2_atom_sequence_append_event(self->port_events_out,
-                                           out_capacity,
-                                           ev);
-        }
+		switch (target)
+		{
+			case 0:
+				lv2_atom_sequence_append_event(self->port_events_out1,
+						out_capacity1,
+						ev);
+				break;
+			case 1:
+				lv2_atom_sequence_append_event(self->port_events_out3,
+						out_capacity3,
+						ev);
+				break;
+		}
     }
+    // Read incoming events
     LV2_ATOM_SEQUENCE_FOREACH(self->port_events_in2, ev)
     {
         if (ev->body.type == self->urid_midiEvent)
@@ -192,13 +236,21 @@ static void run(LV2_Handle instance, uint32_t sample_count)
             }
         }
 
-        if (target_second)
-        {
-            lv2_atom_sequence_append_event(self->port_events_out,
-                                           out_capacity,
-                                           ev);
-        }
+		switch (target)
+		{
+			case 0:
+				lv2_atom_sequence_append_event(self->port_events_out2,
+						out_capacity2,
+						ev);
+				break;
+			case 1:
+				lv2_atom_sequence_append_event(self->port_events_out4,
+						out_capacity4,
+						ev);
+				break;
+		}
     }
+
 }
 
 static void cleanup(LV2_Handle instance)
@@ -207,7 +259,7 @@ static void cleanup(LV2_Handle instance)
 }
 
 static const LV2_Descriptor descriptor = {
-    .URI = "http://moddevices.com/plugins/mod-devel/midi-switchbox-2inx1out",
+    .URI = "http://moddevices.com/plugins/mod-devel/midi-switchbox_1-2_2C",
     .instantiate = instantiate,
     .connect_port = connect_port,
     .activate = activate,
