+++ midi-switchbox_1-2_2C.lv2/midi-switchbox_1-2_2C.ttl
@@ -7,16 +7,16 @@
 @prefix rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
-<http://moddevices.com/plugins/mod-devel/midi-switchbox-1inx3out>
+<http://moddevices.com/plugins/mod-devel/midi-switchbox_1-2_2C>
         a mod:MIDIPlugin ,
             lv2:UtilityPlugin ,
             lv2:Plugin ;
-        doap:name "MIDI SwitchBox 1inx3out" ;
+        doap:name "MIDI SwitchBox 2-1 2C" ;
         doap:license "GPLv2+" ;
         rdfs:comment """
 
-MIDI version of the MOD SwitchBox.
-This switch box receives one MIDI input and channels it through of its three outputs.""" ;
+MIDI version of the MOD Inverted SwitchBox.
+This switch box receives two MIDI inputs and channels it through its output.""" ;
 
         lv2:minorVersion 2 ;
         lv2:microVersion 0 ;
@@ -29,7 +29,7 @@ This switch box receives one MIDI input and channels it through of its three out
                 lv2:name "Target" ;
                 lv2:default 0 ;
                 lv2:minimum 0 ;
-                lv2:maximum 2 ;
+                lv2:maximum 1 ;
                 lv2:portProperty lv2:integer ,
                                  lv2:enumeration ;
                 lv2:scalePoint [
@@ -38,9 +38,6 @@ This switch box receives one MIDI input and channels it through of its three out
                 ] , [
                         rdfs:label "Port 2" ;
                         rdf:value 1 ;
-                ] , [
-                        rdfs:label "Port 3" ;
-                        rdf:value 2 ;
                 ] ;
         ] , [
                 a lv2:InputPort ,
@@ -48,36 +45,52 @@ This switch box receives one MIDI input and channels it through of its three out
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 1 ;
-                lv2:symbol "in" ;
-                lv2:name "In" ;        
+                lv2:symbol "in1" ;
+                lv2:name "In 1" ;
         ] , [
-                a lv2:OutputPort ,
+                a lv2:InputPort ,
                         atom:AtomPort ;
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 2 ;
-                lv2:symbol "out1" ;
-                lv2:name "Out 1" ;
+                lv2:symbol "in2" ;
+                lv2:name "In 2" ;
         ] , [
                 a lv2:OutputPort ,
                         atom:AtomPort ;
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 3 ;
-                lv2:symbol "out2" ;
-                lv2:name "Out 2" ;
+                lv2:symbol "out1" ;
+                lv2:name "out 1" ;
         ] , [
                 a lv2:OutputPort ,
                         atom:AtomPort ;
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 4 ;
+                lv2:symbol "out2" ;
+                lv2:name "out 2" ;
+        ] , [
+                a lv2:OutputPort ,
+                        atom:AtomPort ;
+                atom:bufferType atom:Sequence ;
+                atom:supports midi:MidiEvent ;
+                lv2:index 5 ;
                 lv2:symbol "out3" ;
                 lv2:name "Out 3" ;
+        ] , [
+                a lv2:OutputPort ,
+                        atom:AtomPort ;
+                atom:bufferType atom:Sequence ;
+                atom:supports midi:MidiEvent ;
+                lv2:index 6 ;
+                lv2:symbol "out4" ;
+                lv2:name "Out 4" ;
         ] ;
 
         doap:developer [
-            foaf:name "Jarno Verheesen" ;
+            foaf:name "Jarno Verheesen & Bram Giesen" ;
             foaf:homepage <http://moddevices.com> ;
             foaf:mbox <j.verheesen@hotmail.nl> ;
         ] ;
@@ -89,4 +102,4 @@ This switch box receives one MIDI input and channels it through of its three out
         ] ;
 
         mod:brand "MOD" ;
-        mod:label "MIDI SwitchBox 1inx3out" .
+        mod:label "MIDI InvSwitchBox" .
