+++ midi-switchbox_1-3.lv2/midi-switchbox_1-3.ttl
@@ -7,16 +7,16 @@
 @prefix rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
-<http://moddevices.com/plugins/mod-devel/midi-switchbox-2inx4out>
+<http://moddevices.com/plugins/mod-devel/midi-switchbox_1-3>
         a mod:MIDIPlugin ,
             lv2:UtilityPlugin ,
             lv2:Plugin ;
-        doap:name "MIDI SwitchBox 2inx4out" ;
+        doap:name "MIDI SwitchBox 1inx3out" ;
         doap:license "GPLv2+" ;
         rdfs:comment """
 
-MIDI version of the MOD Inverted SwitchBox.
-This switch box receives two MIDI inputs and channels it through its output.""" ;
+MIDI version of the MOD SwitchBox.
+This switch box receives one MIDI input and channels it through of its three outputs.""" ;
 
         lv2:minorVersion 2 ;
         lv2:microVersion 0 ;
@@ -29,7 +29,7 @@ This switch box receives two MIDI inputs and channels it through its output.""" 
                 lv2:name "Target" ;
                 lv2:default 0 ;
                 lv2:minimum 0 ;
-                lv2:maximum 1 ;
+                lv2:maximum 2 ;
                 lv2:portProperty lv2:integer ,
                                  lv2:enumeration ;
                 lv2:scalePoint [
@@ -38,6 +38,9 @@ This switch box receives two MIDI inputs and channels it through its output.""" 
                 ] , [
                         rdfs:label "Port 2" ;
                         rdf:value 1 ;
+                ] , [
+                        rdfs:label "Port 3" ;
+                        rdf:value 2 ;
                 ] ;
         ] , [
                 a lv2:InputPort ,
@@ -45,52 +48,36 @@ This switch box receives two MIDI inputs and channels it through its output.""" 
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 1 ;
-                lv2:symbol "in1" ;
-                lv2:name "In 1" ;
+                lv2:symbol "in" ;
+                lv2:name "In" ;        
         ] , [
-                a lv2:InputPort ,
+                a lv2:OutputPort ,
                         atom:AtomPort ;
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 2 ;
-                lv2:symbol "in2" ;
-                lv2:name "In 2" ;
+                lv2:symbol "out1" ;
+                lv2:name "Out 1" ;
         ] , [
                 a lv2:OutputPort ,
                         atom:AtomPort ;
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 3 ;
-                lv2:symbol "out1" ;
-                lv2:name "out 1" ;
+                lv2:symbol "out2" ;
+                lv2:name "Out 2" ;
         ] , [
                 a lv2:OutputPort ,
                         atom:AtomPort ;
                 atom:bufferType atom:Sequence ;
                 atom:supports midi:MidiEvent ;
                 lv2:index 4 ;
-                lv2:symbol "out2" ;
-                lv2:name "out 2" ;
-        ] , [
-                a lv2:OutputPort ,
-                        atom:AtomPort ;
-                atom:bufferType atom:Sequence ;
-                atom:supports midi:MidiEvent ;
-                lv2:index 5 ;
                 lv2:symbol "out3" ;
                 lv2:name "Out 3" ;
-        ] , [
-                a lv2:OutputPort ,
-                        atom:AtomPort ;
-                atom:bufferType atom:Sequence ;
-                atom:supports midi:MidiEvent ;
-                lv2:index 6 ;
-                lv2:symbol "out4" ;
-                lv2:name "Out 4" ;
         ] ;
 
         doap:developer [
-            foaf:name "Jarno Verheesen & Bram Giesen" ;
+            foaf:name "Jarno Verheesen" ;
             foaf:homepage <http://moddevices.com> ;
             foaf:mbox <j.verheesen@hotmail.nl> ;
         ] ;
@@ -102,4 +89,4 @@ This switch box receives two MIDI inputs and channels it through its output.""" 
         ] ;
 
         mod:brand "MOD" ;
-        mod:label "MIDI InvSwitchBox" .
+        mod:label "MIDI SwitchBox 1inx3out" .
