+++ midi-switchbox_2-1.lv2/manifest.ttl
@@ -1,0 +1,7 @@
+@prefix lv2:  <http://lv2plug.in/ns/lv2core#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+
+<http://moddevices.com/plugins/mod-devel/midi-switchbox_2-1>
+    a lv2:Plugin ;
+    lv2:binary <midi-switchbox_2-1.so>  ;
+    rdfs:seeAlso <midi-switchbox_2-1.ttl> , <modgui.ttl> .
