+++ midi-switchbox_2-1.lv2/midi-switchbox_2-1.ttl
@@ -7,7 +7,7 @@
 @prefix rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
 @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
 
-<http://moddevices.com/plugins/mod-devel/midi-switchbox-2inx1out>
+<http://moddevices.com/plugins/mod-devel/midi-switchbox_2-1>
         a mod:MIDIPlugin ,
             lv2:UtilityPlugin ,
             lv2:Plugin ;
